Make Calypso generate two builds: evergreen and fallback.
We currently have a single, fallback build, which includes support
for every supported browser. This results in a large amount of
polyfills and syntactic transformations which bloat the bundle and
increase browser download and execution times.

This PR aims to generate a separate evergreen build in production,
and serve that build instead to evergreen browsers, by doing
user agent detection. Other browsers will still work, since they
have the fallback bundle, but known modern browsers will get the
benefit of a smaller bundle.